UPBGE: Start game engine only from first wmWindow
If there are several windows with several 3D views in a .blend, we could eventually
try to use the right window and the right context BUT:

When we load another .blend during bge runtime, it would cause a mess to
choose the right window and obtain the right bContext.

Imo, it's easier to always force to use first window both in standalone
and embedded.

It's to prevent such crashes: https://github.com/UPBGE/upbge/issues/1659

disallowing start from another window